requirements.txt
aiobotocore==2.12.3
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
bayesian-optimization==1.4.3
botocore==1.34.69
catboost==1.2.5
certifi==2024.2.2
charset-normalizer==3.3.2
chart-studio==1.1.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cramjam==2.8.3
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
Faker==25.0.1
fastjsonschema==2.19.1
fastparquet==2024.2.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
graphviz==0.20.3
hyperopt==0.2.7
idna==3.7
imbalanced-learn==0.12.2
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
kneed==0.8.5
lightgbm==4.3.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
missingno==0.5.2
multidict==6.0.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==16.0.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
retrying==1.3.4
rpds-py==0.18.1
s3fs==2024.3.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
tenacity==8.3.0
threadpoolctl==3.5.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
wordcloud==1.9.3
wrapt==1.16.0
xgboost==2.0.3
yarl==1.9.4
yellowbrick==1.5
